requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-ppc64le
_libgcc_mutex=0.1=main
_py-xgboost-mutex=1.0=gpu_610.g99281e0
_pytorch_select=2.0=gpu_20258.g46cbccf
_tflow_select=2.1.0=gpu_882.g60d75b8
absl-py=0.7.1=py37_0
apex=0.1.0_1.6.2=py37_614.g6ced8ce
astor=0.7.1=py37_0
atomicwrites=1.4.0=py_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=openblas
bokeh=1.4.0=py37_0
boost=1.67.0=py37_4
brotlipy=0.7.0=py37h140841e_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.17.1=h140841e_0
ca-certificates=2021.7.5=h6ffa863_1
caffe=1.0_1.6.2=5202.g79d24cd
caffe-base=1.0_1.6.2=gpu_py37_5202.g79d24cd
cairo=1.16.0=he491a88_1
certifi=2021.5.30=py37h6ffa863_0
cffi=1.14.6=py37hf9d8e4b_0
chardet=3.0.4=py37h6ffa863_1003
click=7.0=py37_0
cloudpickle=1.6.0=py_0
coverage=4.5.4=py37h7b6447c_0
cryptography=3.4.7=py37h7ed74fa_0
cudatoolkit=10.1.243=616.gc122b8b
cudnn=7.6.3_10.1=590.g5627c5e
cycler=0.10.0=py37_0
cytoolz=0.11.0=py37h7b6447c_0
dask=2.3.0=py_0
dask-core=2.3.0=py_0
dask-cuda=0.9.1=py37_592.gdc4b434
dask-xgboost=0.1.7=py37_597.g7547b0f
ddl=1.5.0=py37_1306.g2d85fa8
ddl-tensorflow=1.5.0=py37_1025.g27a3898
decorator=5.0.9=pyhd3eb1b0_0
distributed=2.3.2=py_1
ffmpeg=4.2.2=h20bf706_0
fontconfig=2.13.1=ha0a49a9_0
freeglut=3.0.0=hf484d3e_5
freetype=2.10.4=h5ab3b9f_0
fsspec=2021.7.0=pyhd3eb1b0_0
future=0.17.1=py37_0
gast=0.2.2=py37_0
gflags=2.2.2=1642.g3029267
glib=2.69.0=h1318424_0
glog=0.3.5=1631.ge7b0425
gmp=6.2.1=h29c3540_0
gnutls=3.6.15=hd39c10c_0
google-pasta=0.1.6=py37_584.g6a97e08
graphite2=1.3.14=h23475e2_0
graphsurgeon=0.4.1=py37_634.g4ac44fb
grpcio=1.16.1=py37hf8bcb03_1
h5py=2.8.0=py37h8d01980_0
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
heapdict=1.0.1=py_0
hypothesis=3.59.1=py37h39e3cac_0
icu=58.2=he6710b0_3
idna=2.8=py37_0
imageio=2.9.0=pyhd3eb1b0_0
importlib-metadata=3.10.0=py37h6ffa863_0
importlib_metadata=3.10.0=hd3eb1b0_0
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.26.0=py37he95b402_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jasper=2.0.14=h07fcdf6_1
jedi=0.18.0=py37h6ffa863_1
jinja2=3.0.1=pyhd3eb1b0_0
joblib=0.13.2=py37_0
jpeg=9b=hcb7ba68_2
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.7.1=py37h6ffa863_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.0=py_1
kiwisolver=1.3.1=py37h29c3540_0
lame=3.100=h7b6447c_0
ld_impl_linux-ppc64le=2.33.1=h0f24833_7
leveldb=1.20=hf484d3e_1
libboost=1.67.0=h46d08c1_4
libffi=3.3=he6710b0_2
libgcc-ng=8.2.0=h822a55f_1
libgfortran-ng=7.3.0=h822a55f_1
libglu=9.0.0=hf484d3e_1
libidn2=2.3.2=h140841e_0
libopenblas=0.3.6=h5a2b251_1
libopencv=3.4.7=748.g5e2fea8
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.8.0=596.gf044ff6
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=8.2.0=h822a55f_1
libtasn1=4.16.0=h140841e_0
libtiff=4.2.0=h781710b_0
libunistring=0.9.10=h140841e_0
libuuid=1.0.3=h1bed415_2
libvpx=1.7.0=hf484d3e_0
libwebp-base=1.2.0=h140841e_0
libxcb=1.14=h7b6447c_0
libxgboost-base=0.90=gpu_610.g99281e0
libxml2=2.9.12=hcd9f32a_0
llvmlite=0.29.0=py37hd408876_0
lmdb=0.9.22=hf484d3e_1
locket=0.2.1=py37h6ffa863_1
lz4-c=1.9.3=h29c3540_0
markdown=3.1.1=py37_0
markupsafe=2.0.1=py37h140841e_0
matplotlib=3.3.4=py37h6ffa863_0
matplotlib-base=3.3.4=py37he087750_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mock=2.0.0=py37_0
more-itertools=8.8.0=pyhd3eb1b0_0
msgpack-python=1.0.2=py37h66086b3_1
nccl=2.4.8=586.gdba67b7
ncurses=6.2=he6710b0_1
nettle=3.7.3=hdc176a3_1
networkx=2.2=py37_1
ninja=1.9.0=py37hfd86e86_0
nomkl=3.0=0
numactl=2.0.12=591.g5f5b313
numba=0.45.1=py37h962f231_0
numpy=1.16.6=py37hece86af_3
numpy-base=1.16.6=py37h0a9401a_3
olefile=0.46=py37_0
onnx=1.5.0=py37_632.g66eb6ad
openblas=0.3.6=1
openblas-devel=0.3.6=1
opencv=3.4.7=py37_748.g5e2fea8
openh264=2.1.0=hd408876_0
openssl=1.1.1k=h140841e_0
packaging=21.0=pyhd3eb1b0_0
pai4sk=1.5.0=py37_1086.g5abf42e
pandas=1.2.4=py37haab0e66_0
parso=0.8.2=pyhd3eb1b0_0
partd=1.2.0=pyhd3eb1b0_0
pbr=5.6.0=pyhd3eb1b0_0
pciutils=3.6.2=589.g384a96e
pcre=8.45=h29c3540_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=7.1.2=py37haac5956_0
pip=21.1.3=py37h6ffa863_0
pixman=0.40.0=h7b6447c_0
pluggy=0.13.1=py37h6ffa863_0
powerai=1.6.2=633.g69d9f93
powerai-license=1.6.2=734.g398a206
powerai-release=1.6.2=590.g8b407ea
powerai-tools=1.6.2=584.ga57254d
prompt-toolkit=3.0.17=pyh06a4308_0
protobuf=3.8.0=py37_606.ge232d90
psutil=5.5.0=py37h7b6447c_0
ptyprocess=0.7.0=pyhd3eb1b0_2
py=1.10.0=pyhd3eb1b0_0
py-boost=1.67.0=py37h04863e7_4
py-opencv=3.4.7=py37_748.g5e2fea8
py-xgboost-base=0.90=gpu_py37_610.g99281e0
py-xgboost-gpu=0.90=610.g99281e0
pycparser=2.20=py_2
pygments=2.10.0=pyhd3eb1b0_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pysocks=1.7.1=py37_1
pytest=4.4.2=py37_0
python=3.7.10=h836d2c2_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-lmdb=0.94=py37h14c3975_0
pytorch=1.2.0=20258.g46cbccf
pytorch-base=1.2.0=gpu_py37_20258.g46cbccf
pytz=2021.1=pyhd3eb1b0_0
pywavelets=1.1.1=py37h7b6447c_2
pyyaml=5.4.1=py37h140841e_1
pyzmq=22.2.1=py37h29c3540_1
readline=8.1=h140841e_0
requests=2.22.0=py37_1
scikit-image=0.15.0=py37he6710b0_0
scikit-learn=0.22.1=py37h22eb022_0
scipy=1.3.1=py37he2b7bc3_0
setuptools=52.0.0=py37h6ffa863_0
simsearch=1.1.0=py37_783.g804cfc8
six=1.12.0=py37_0
snapml-spark=1.4.0=py37_960.gb5ce883
snappy=1.1.7=h1532aa0_3
sortedcontainers=2.4.0=pyhd3eb1b0_0
spectrum-mpi=10.03=640.g989d89f
sqlite=3.36.0=hd7247d8_0
tabulate=0.8.2=py37_0
tbb=2021.2.0=h66086b3_0
tblib=1.7.0=py_0
tensorboard=1.15.0=py37_ab7f72a_3676.g60d75b8
tensorflow=1.15.5=gpu_py37_882.g60d75b8
tensorflow-base=1.15.5=gpu_py37_3db52be_64480.g405b235
tensorflow-estimator=1.15.2=py37_25220bb_1398.g60d75b8
tensorflow-gpu=1.15.5=882.g60d75b8
tensorflow-large-model-support=2.0.2=py37_989.gbbbbc51
tensorflow-probability=0.8.0=py37_b959b26_2728.g60d75b8
tensorflow-serving-api=1.15.0=py37_748217e_5112.gd2f86b0
tensorrt=6.0.1.5=py37_634.g4ac44fb
termcolor=1.1.0=py37h6ffa863_1
tf_cnn_benchmarks=1.15=gpu_py37_1392.g0ecd171
tk=8.6.10=hbc83047_0
toolz=0.11.1=pyhd3eb1b0_0
torchtext=0.4.0=py37_596.g730653b
torchvision-base=0.4.0=gpu_py37_613.g2f2b557
tornado=6.1=py37h140841e_0
tqdm=4.32.1=py_0
traitlets=5.0.5=pyhd3eb1b0_0
typing=3.6.4=py37_0
typing_extensions=3.7.2=py37_1596.g5249206
uff=0.6.5=py37_634.g4ac44fb
urllib3=1.25.11=py_0
wcwidth=0.2.5=py_0
werkzeug=0.15.4=py_0
wheel=0.36.2=pyhd3eb1b0_0
wrapt=1.11.2=py37h7b6447c_0
x264=1!157.20191217=h7b6447c_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zeromq=4.3.4=h29c3540_0
zict=2.0.0=pyhd3eb1b0_0
zipp=3.5.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=hc52992f_0
